feat: nest model routes within `models/` folder
