+++ src/pages/models/[model]/index.tsx
