+++ src/pages/models/index.tsx
@@ -1,0 +1,7 @@
+import { ModelsLayout, PageWithLayout } from '@/layouts';
+
+const ModelsHome: PageWithLayout = () => {
+  return <div>Home</div>;
+};
+ModelsHome.layout = ModelsLayout;
+export default ModelsHome;
